First prototype
This is the first draft implementation. We currently use sqlite to
create an in memory database, `rg` to get relevant metadata lines on
management files, parse those lines, populate the db with time tracking
data, and print some meaningful queries. In the future we'll have other
features, but this should be enough for a proof of concept and real life
usage/testing.

This is my first project using crystal and I didn't take the time to set
linters and other best practices. The structure might change as well,
but it is a starting point nonetheless.